Extract deprecated param logger to static function
... and remove deprecated class AmountParser.
