+++ src/Infrastructure/Validation/NullDomainNameValidator.php
@@ -2,7 +2,7 @@
 
 declare( strict_types = 1 );
 
-namespace WMDE\Fundraising\Frontend\Infrastructure;
+namespace WMDE\Fundraising\Frontend\Infrastructure\Validation;
 
 use WMDE\FunValidators\DomainNameValidator;
